MVR-304 Add pinned recipes feature and tests
This commit introduces a new functionality where users can pin their favorite recipes. This includes a new 'Pinned' route, corresponding page and components, and tests. It implements the pin/unpin functionality by making use of local storage. Included are also UI elements such as a new 'Pinned' link in the header and an icon/button for pinning and unpinning recipes from the recipe overview and details.