MVR-304 Refactor rendering steps in Pinned and test scripts (ESLINT)
The 'act' wrapper in the rendering step for 'Pinned' component in the tests has been removed for simplicity. Also, the 'within' import in the tests scripts has been adjusted to come from '@testing-library/react' instead of '@testing-library/dom'. Furthermore, 'baseUrl' has been added as a dependency for effect in the 'Pinned' component for better re-rendering control.